requirements.txt
sklearn
numpy
scipy
gensim
google_compute_engine
flask
flask-jsonpify
flask-sqlalchemy
flask-restful
flask-cors
google-python-cloud-debugger
google-cloud-storage
